Fix timer delaying after some time running
After a day or 2 of running, the timer would start to delay too much.
Possibly because of conversions between different types of variables.
So I've changed all types to uint64_t to have the same size as the total
of what comes from the IOB_TIMER_GET_DATA_HIGH()/LOW() functions.